Document how to change the CLI editor used by geordi
I was surprised that `geordi drop-databases` uses `nano`, while my `$EDITOR` is set to `vim`.

Looking at geordis source code did not help me at first, but I found the solution to my problem in a PR description. With this patch it should be slightly easier for the next person.